morphos/services/backend-service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.0
aiohttp
opencv-python-headless==4.8.1.78
numpy==1.24.4
pydantic
pydantic-settings>=2.0.0
requests
python-jose==3.3.0
python-dotenv==0.21.0
httpx==0.23.0
pymongo==4.3.3
python-dotenv==0.21.0
email-validator==2.0.0
pymongo==4.3.3